Update token label component
