+++ src/shared/ui/components/token-label/token-label.scss
