+++ src/shared/ui/components/token-label/token-label.tsx
@@ -1,0 +1,28 @@
+import React from "react";
+
+import { NearGenericTokenIconFilled, NearIconFilled } from "../icons";
+import "./token-label.scss";
+
+export const TokenLabel = ({
+    icon = <NearGenericTokenIconFilled />,
+    native,
+    symbol,
+}: {
+    icon?: string | JSX.Element | null;
+    native?: boolean;
+    symbol?: string;
+}) => (
+    <span className="token-label">
+        {!native && typeof icon === "string" ? (
+            <img
+                className="icon"
+                loading="lazy"
+                src={icon}
+            />
+        ) : (
+            <span className="icon">{native ? <NearIconFilled /> : icon}</span>
+        )}
+
+        <span className="symbol">{native ? "NEAR" : symbol}</span>
+    </span>
+);
